requirements.txt
streamlit
requests
python-dotenv
tldextract
pandas
numpy
sentence-transformers
torch

